requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.4
djangorestframework==3.8.2
idna==2.6
pytz==2018.3
PyYAML==3.12
requests==2.18.4
uber-rides==0.6.0
urllib3==1.22
